+++ src/main/java/CDAccount.java
@@ -1,5 +1,5 @@
 public class CDAccount extends Account {
-    public CDAccount(int amount, double apr) {
+    public CDAccount(double amount, double apr) {
         this.amount = amount;
         this.apr = apr;
         this.maxDeposit = -1;
